\docType{package}
\name{bmatrix}
\alias{bmatrix}
\alias{bmatrix-package}
\alias{package-bmatrix}
\title{Block matrix}
\usage{
  bmatrix(data, nrow = 1, ncol = 1, byrow = FALSE)
}
\arguments{
  \item{data}{A list of matrices.}

  \item{nrow}{Number of block rows.}

  \item{ncol}{Number of block columns.}

  \item{byrow}{Fill matrix by block rows?}
}
\value{
  A matrix
}
\description{
  Create matrices out of blocks.

  Create a matrix out of blocks.
}
\author{
  Steve Walker
}

